move map UI into feature package
